back/requirements.txt
Django==3.2.9
django-cors-headers==3.10.0
djangorestframework==3.12.4
djangorestframework-jwt==1.11.0
djongo==1.3.6
dnspython==2.1.0
pytz==2021.3